requirements.txt
flask
numpy
pandas
scikit-learn
tensorflow
yfinance
nltk
transformers
newsapi-python
tweepy
alpha_vantage